requirements.txt
numpy
opencv-python
streamlit
mysql-connector-python